# Conflicts:
#	src/Arch/Mos6502/Mos6502ProcessorArchitecture.cs
#	src/Core/Core.csproj
#	src/Decompiler/Decompiler.cs

+++ src/Core/Expressions/Constant.cs
@@ -860,7 +860,7 @@ namespace Reko.Core.Expressions
 
         public override long ToInt64()
         {
-            return value;
+            return (int)value;
         }
     }
